// src/models/buyer.ts
export interface Buyer {
  name: string;
  email: string;
  role: string;
  position: string;
  taxid: string;
  typeofbusiness: string;
  businesscontact: number;
  weekstart: string;
  profilepic: string;
  opentime: string;
  fax: string;
  experience: string;
  department: string;
  address: string;
  business: string;
  closedtime: string;
  companylogo: string;
  contactno: number;
  companywebsite: string;
  weekend:string;
  rep:string[];
  date:Date;
  password:string;
  street:string;
  city:string;
}